+++ src/main/java/org/neo4j/example/unmanagedextension/MyService.java
@@ -83,7 +83,7 @@ public class MyService {
 			} catch (IOException e1) {
 				neo4jLogger.logMessage("AAPN is not accessible\n"+e1.toString());
 			}
-			long jobID = (int) jsonMap.get(JsonKeyConstants.JOB_ID);
+			long jobID = (long) jsonMap.get(JsonKeyConstants.JOB_ID);
 			try {
 				h2Helper.updateJobWithResults(jobID, aapn+" - jobID:" + jobID + " CAKILDI ");
 			} catch (SQLException e1) {
